Consume zonke package directly in main
