--- packages/hosting/package.json
@@ -1,37 +1,0 @@
-{
-    "name": "@onlook/hosting",
-    "description": "A hosting library for Onlook",
-    "main": "./src/index.ts",
-    "type": "module",
-    "module": "src/index.ts",
-    "types": "src/index.ts",
-    "version": "0.0.0",
-    "private": true,
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/onlook-dev/onlook.git"
-    },
-    "scripts": {
-        "clean": "rm -rf node_modules",
-        "lint": "eslint --fix .",
-        "format": "prettier --write .",
-        "typecheck": "tsc --noEmit",
-        "test": "bun test"
-    },
-    "keywords": [
-        "onlook",
-        "hosting"
-    ],
-    "author": {
-        "name": "Onlook",
-        "email": "contact@onlook.dev"
-    },
-    "license": "Apache-2.0",
-    "homepage": "https://onlook.dev",
-    "devDependencies": {
-        "@onlook/typescript": "*"
-    },
-    "dependencies": {
-        "@zonke-cloud/sdk": "^0.1.2"
-    }
-}
