--- packages/hosting/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "extends": "@onlook/typescript/base.json",
-    "compilerOptions": {
-        "baseUrl": "."
-    },
-    "include": ["src", "test"],
-    "exclude": ["node_modules"]
-}
